+++ projects/ngx-loader/src/lib/spinner-loader/spinner-loader.component.scss
@@ -21,7 +21,7 @@
   display: block;
   width: 25%;
   height: 25%;
-  background-color: #fff;
+  background-color: #e4b2b2;
   border-radius: 100%;
   animation: sk-chase-dot-before 2.0s infinite ease-in-out both;
 }
