--- projects__/demo/src/app/app.component.scss
@@ -1,11 +1,0 @@
-.box{
-    width: 200px;
-    height: auto;
-    background-color: aliceblue;
-    margin: 0.5rem;
-    padding: 1rem;
-    display: flex;
-    justify-content: center;
-    align-content: center;
-    align-items: center;
-}
