--- projects__/demo/src/app/app.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
-})
-export class AppComponent {
-
-    public loading: boolean = false;
-    load() {
-        this.loading = !this.loading
-
-    }
-}
