--- projects__/demo/src/styles.scss
@@ -1,2 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-@import "node_modules/bootstrap/scss/bootstrap.scss";
