--- projects__/ngx-loader/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@ildug/ngx-loader",
-  "version": "1.0.0",
-  "author": {
-    "name": "ilDug"
-  },
-  "description": "spinner loader angular directive",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ilDug/ngx-loader.git"
-  },
-  "license": "MIT",
-  "peerDependencies": {
-    "@angular/common": ">12",
-    "@angular/core": ">12"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  }
-}
