--- projects__/ngx-loader/src/lib/bouncing-loader/bouncing-loader.component.html
@@ -1,9 +1,0 @@
-<div class="spinner-container">
-    <div class="spinner">
-        <div [ngStyle]="{'background-color': color}" class="rect1"></div>
-        <div [ngStyle]="{'background-color': color}" class="rect2"></div>
-        <div [ngStyle]="{'background-color': color}" class="rect3"></div>
-        <div [ngStyle]="{'background-color': color}" class="rect4"></div>
-        <div [ngStyle]="{'background-color': color}" class="rect5"></div>
-    </div>
-</div>
