--- projects__/ngx-loader/src/lib/bouncing-loader/bouncing-loader.component.ts
@@ -1,11 +1,0 @@
-import { Component, HostBinding, OnInit } from '@angular/core';
-import { LoaderComponent } from '../loader.component';
-
-@Component({
-  selector: 'dag-bouncing-loader',
-  templateUrl: './bouncing-loader.component.html',
-  styleUrls: ['./bouncing-loader.component.scss']
-})
-export class BouncingLoaderComponent extends LoaderComponent {
-    @HostBinding('class.full-screen') fullScreen: boolean;
-}
