--- projects__/ngx-loader/src/lib/loader.component.ts
@@ -1,7 +1,0 @@
-export abstract class LoaderComponent {
-   public color: string = '#333';
-   public overlayPosition: 'absolute' | 'fixed' = "absolute"
-    public width: number;
-   public height: number;
-   public fullScreen: boolean;
-}
