--- projects__/ngx-loader/src/lib/pulsing-loader/pulsing-loader.component.html
@@ -1,6 +1,0 @@
-<div class="spinner-container">
-    <div class="spinner">
-        <div class="double-bounce1" [ngStyle]="{'background-color': color}"></div>
-        <div class="double-bounce2" [ngStyle]="{'background-color': color}"></div>
-        </div>
-</div>
