--- projects__/ngx-loader/src/lib/pulsing-loader/pulsing-loader.component.ts
@@ -1,11 +1,0 @@
-import { Component, HostBinding, OnInit } from '@angular/core';
-import { LoaderComponent } from '../loader.component';
-
-@Component({
-  selector: 'dag-pulsing-loader',
-  templateUrl: './pulsing-loader.component.html',
-  styleUrls: ['./pulsing-loader.component.scss']
-})
-export class PulsingLoaderComponent extends LoaderComponent {
-    @HostBinding('class.full-screen') fullScreen: boolean;
-}
