--- projects__/ngx-loader/src/lib/spinner-loader/spinner-loader.component.html
@@ -1,10 +1,0 @@
-<div class="spinner-container">
-    <div class="sk-chase">
-        <div class="sk-chase-dot"></div>
-        <div class="sk-chase-dot"></div>
-        <div class="sk-chase-dot"></div>
-        <div class="sk-chase-dot"></div>
-        <div class="sk-chase-dot"></div>
-        <div class="sk-chase-dot"></div>
-    </div>
-</div>
