--- projects__/ngx-loader/src/lib/styles.scss
@@ -1,42 +1,0 @@
-.spinner-container{
-    display: flex;
-    justify-content: center;
-    align-content: center;
-    align-items: center;
-    height: 100%;
-    width: 100%;
-    
-}
-
-.spinner-container{
-    // position: relative;
-    position: absolute;
-    top: 0;
-    left: 0;
-
-    &::before{
-        content: ""; // ::before and ::after both require content
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        // background-image: linear-gradient(120deg, #ffbffc, rgb(51, 163, 255));
-        background-color: #ccc;
-        opacity: .7;
-        z-index: 1;
-    }
-
-    & > *{
-        z-index: 99;
-    }
-}
-
-
-
-:host.full-screen .spinner-container{
-    position: fixed !important;
-    &::before{
-        position: fixed !important;
-    }
-}
