--- projects__/ngx-loader/src/public-api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of ngx-loader
- */
-
-export * from './lib/ngx-loader.module';
-export * from './lib/loader.directive';
